#578: Fixed support for generating RTl PDFs GlobalDigitalLibraryio/issues#578
