+++ build.sbt
@@ -71,6 +71,7 @@ lazy val book_api = (project in file(".")).
       "com.openhtmltopdf" % "openhtmltopdf-core" % OpenHtmlPdfVersion,
       "com.openhtmltopdf" % "openhtmltopdf-pdfbox" % OpenHtmlPdfVersion,
       "com.openhtmltopdf" % "openhtmltopdf-jsoup-dom-converter" % OpenHtmlPdfVersion,
+      "com.openhtmltopdf" % "openhtmltopdf-rtl-support" % OpenHtmlPdfVersion,
       "com.github.blemale" %% "scaffeine" % "2.3.0",
       "commons-validator" % "commons-validator" % "1.6",
       "io.sentry" % "sentry-log4j2" % "1.7.3"
