+++ src/main/scala/io/digitallibrary/bookapi/service/PdfService.scala
@@ -10,18 +10,15 @@ package io.digitallibrary.bookapi.service
 import java.io.{ByteArrayInputStream, ByteArrayOutputStream, InputStream}
 
 import com.amazonaws.services.s3.model.{GetObjectRequest, ObjectMetadata, PutObjectRequest, S3Object}
-import com.openhtmltopdf.extend.FSSupplier
 import com.openhtmltopdf.pdfboxout.PdfRendererBuilder
 import com.typesafe.scalalogging.LazyLogging
 import io.digitallibrary.bookapi.BookApiProperties
 import io.digitallibrary.bookapi.integration.AmazonClient
-import io.digitallibrary.bookapi.model.api.{Chapter, NotFoundException, PdfStream}
-import io.digitallibrary.bookapi.model.domain.{BookFormat, PdfCss}
+import io.digitallibrary.bookapi.model.api.{NotFoundException, PdfStream}
+import io.digitallibrary.bookapi.model.domain.BookFormat
 import io.digitallibrary.bookapi.repository.{BookRepository, TranslationRepository}
+import io.digitallibrary.bookapi.service.pdf.Pdf
 import io.digitallibrary.language.model.LanguageTag
-import org.jsoup.Jsoup
-import org.jsoup.nodes.Document
-import org.jsoup.nodes.Entities.EscapeMode
 
 import scala.util.{Failure, Success, Try}
 
@@ -46,35 +43,6 @@ trait PdfService {
       }
     }
 
-    case class FontDefinition(fontFile: String, fontName: String)
-
-    class NotoFontSupplier(stream: InputStream) extends FSSupplier[InputStream] {
-      override def supply(): InputStream = stream
-    }
-
-    val DefaultFont = FontDefinition("/NotoSans-Regular.ttf", "Noto Sans")
-    val fonts = Map(
-      "ethiopic" -> FontDefinition("/NotoSansEthiopic.ttf", "Noto Sans Ethiopic"),
-      "devangari" -> FontDefinition("/NotoSansDevanagari-Regular.ttf", "Noto Sans Devanagari"),
-      "bengali" -> FontDefinition("/NotoSansBengali-Regular.ttf", "Noto Sans Bengali"),
-      "khmer" -> FontDefinition("/NotoSansKhmer-Regular.ttf", "Noto Sans Khmer"),
-      "tamil" -> FontDefinition("/NotoSansTamil-Regular.ttf", "Noto Sans Tamil"),
-      "arabic" -> FontDefinition("/NotoSansArabic-Regular.ttf", "Noto Sans Arabic")
-    )
-    val fontDefinitions = Map(
-      LanguageTag("tir-et") -> fonts("ethiopic"),
-      LanguageTag("tir") -> fonts("ethiopic"),
-      LanguageTag("amh") -> fonts("ethiopic"),
-      LanguageTag("mar") -> fonts("devangari"),
-      LanguageTag("hin") -> fonts("devangari"),
-      LanguageTag("ben") -> fonts("bengali"),
-      LanguageTag("nep") -> fonts("devangari"),
-      LanguageTag("khm") -> fonts("khmer"),
-      LanguageTag("tam") -> fonts("tamil"),
-      LanguageTag("ar-ae") -> fonts("arabic"),
-      LanguageTag("ar") -> fonts("arabic")
-    )
-
     def getPdf(language: LanguageTag, uuid: String): Option[PdfStream] = {
       unFlaggedTranslationsRepository.withUuId(uuid) match {
         case Some(translation) => translation.bookFormat match {
@@ -91,58 +59,13 @@ trait PdfService {
 
     def createPdf(language: LanguageTag, uuid: String): Option[PdfRendererBuilder] = {
       unFlaggedTranslationsRepository.withUuId(uuid).map(translation => {
-        val source = bookRepository.withId(translation.bookId).map(_.source)
-
-        val chapters = readService.chaptersForIdAndLanguage(translation.bookId, language).flatMap(ch => readService.chapterForBookWithLanguageAndId(translation.bookId, language, ch.id))
-        val preprocessedChapters = preprocessChapters(source, chapters)
-        val fonts = fontDefinitions.get(language).toSeq :+ DefaultFont
-
-        val bookAsHtml =
-          s"""
-             |<html>
-             |  <head>
-             |    <style language='text/css'>${PdfCss(source, translation.pageOrientation, fonts.map(_.fontName)).asString}</style>
-             |  </head>
-             |  <body>
-             |    ${preprocessedChapters.zipWithIndex.map { case (c, i) => s"<div class='page page_$i ${c.chapterType.toString.toLowerCase}'>${c.content}</div>" }.mkString("\n")}
-             |  </body>
-             |</html>
-           """.stripMargin
-
-        val rendererBuilder = new PdfRendererBuilder().withHtmlContent(bookAsHtml, "/")
-        fonts.foldLeft(rendererBuilder) { (builder, font) =>
-          builder.useFont(new NotoFontSupplier(getClass.getResourceAsStream(font.fontFile)), font.fontName)
-        }
+        Pdf(language=language,
+          chapters=readService.chaptersForIdAndLanguage(translation.bookId, language).flatMap(ch => readService.chapterForBookWithLanguageAndId(translation.bookId, language, ch.id)),
+          translation=translation,
+          source=bookRepository.withId(translation.bookId).map(_.source))
+          .create()
       })
     }
-
-    private def preprocessChapters(source: Option[String], chapters: Seq[Chapter]) = {
-      def processStoryWeaverFirstPage(c: Chapter): Chapter = {
-        val document = Jsoup.parseBodyFragment(c.content)
-        val images = document.select("img")
-        for (i <- 0 until images.size()) {
-          val image = images.get(i)
-          if (i == 0) { // First image
-            image.addClass("cover")
-          } else {
-            image.addClass("logo")
-          }
-        }
-        document.outputSettings().syntax(Document.OutputSettings.Syntax.xml)
-        document.outputSettings().escapeMode(EscapeMode.xhtml).prettyPrint(false)
-        c.copy(content = document.select("body").html())
-      }
-
-      source match {
-        case Some("storyweaver") =>
-          chapters match {
-            case first :: rest => processStoryWeaverFirstPage(first) :: rest
-            case _ => chapters
-          }
-        case _ => chapters
-      }
-    }
-
 
     def getFromS3(uuid: String): Try[S3Object] = {
       Try(amazonClient.getObject(new GetObjectRequest(BookApiProperties.StorageName, s"$uuid.pdf"))) match {
